pyproject.toml
[tool.poetry]
name = "bidbybid"
version = "0.1.0"
description = "Scrape ebay sold auctions"
authors = ["Leonardo Cavaletti <impeto.blu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
colorama = "^0.4.4"
selenium = "^3.141.0"
price-parser = "^0.3.4"
chromedriver-autoinstaller = "^0.2.2"
python-dateutil = "^2.8.1"
pygal = "^2.4.0"
lxml = "^4.6.2"

[tool.poetry.dev-dependencies]
taskipy = "^1.6.0"
black = "^20.8b1"

[tool.taskipy.tasks]
start = "python bidbybid.py"
freeze = "poetry export -f requirements.txt > requirements.txt --without-hashes"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
